Split map_entry tests into fixable and unfixable
