+++ tests/ui/entry_unfixable.stderr
@@ -1,43 +1,16 @@
 error: usage of `contains_key` followed by `insert` on a `HashMap`
-  --> $DIR/entry.rs:10:5
+  --> $DIR/entry_unfixable.rs:10:5
    |
 LL | /     if !m.contains_key(&k) {
-LL | |         m.insert(k, v);
-LL | |     }
-   | |_____^ help: consider using: `m.entry(k).or_insert(v)`
+LL | |         m.insert(k, v)
+LL | |     } else {
+LL | |         None
+LL | |     };
+   | |_____^
    |
    = note: `-D clippy::map-entry` implied by `-D warnings`
-
-error: usage of `contains_key` followed by `insert` on a `HashMap`
-  --> $DIR/entry.rs:16:5
-   |
-LL | /     if !m.contains_key(&k) {
-LL | |         foo();
-LL | |         m.insert(k, v);
-LL | |     }
-   | |_____^
-   |
 help: consider using `m.entry(k)`
-  --> $DIR/entry.rs:16:5
-   |
-LL | /     if !m.contains_key(&k) {
-LL | |         foo();
-LL | |         m.insert(k, v);
-LL | |     }
-   | |_____^
-
-error: usage of `contains_key` followed by `insert` on a `HashMap`
-  --> $DIR/entry.rs:23:5
-   |
-LL | /     if !m.contains_key(&k) {
-LL | |         m.insert(k, v)
-LL | |     } else {
-LL | |         None
-LL | |     };
-   | |_____^
-   |
-help: consider using `m.entry(k)`
-  --> $DIR/entry.rs:23:5
+  --> $DIR/entry_unfixable.rs:10:5
    |
 LL | /     if !m.contains_key(&k) {
 LL | |         m.insert(k, v)
@@ -47,7 +20,7 @@ LL | |     };
    | |_____^
 
 error: usage of `contains_key` followed by `insert` on a `HashMap`
-  --> $DIR/entry.rs:31:5
+  --> $DIR/entry_unfixable.rs:18:5
    |
 LL | /     if m.contains_key(&k) {
 LL | |         None
@@ -57,7 +30,7 @@ LL | |     };
    | |_____^
    |
 help: consider using `m.entry(k)`
-  --> $DIR/entry.rs:31:5
+  --> $DIR/entry_unfixable.rs:18:5
    |
 LL | /     if m.contains_key(&k) {
 LL | |         None
@@ -67,7 +40,7 @@ LL | |     };
    | |_____^
 
 error: usage of `contains_key` followed by `insert` on a `HashMap`
-  --> $DIR/entry.rs:39:5
+  --> $DIR/entry_unfixable.rs:26:5
    |
 LL | /     if !m.contains_key(&k) {
 LL | |         foo();
@@ -78,7 +51,7 @@ LL | |     };
    | |_____^
    |
 help: consider using `m.entry(k)`
-  --> $DIR/entry.rs:39:5
+  --> $DIR/entry_unfixable.rs:26:5
    |
 LL | /     if !m.contains_key(&k) {
 LL | |         foo();
@@ -89,7 +62,7 @@ LL | |     };
    | |_____^
 
 error: usage of `contains_key` followed by `insert` on a `HashMap`
-  --> $DIR/entry.rs:48:5
+  --> $DIR/entry_unfixable.rs:35:5
    |
 LL | /     if m.contains_key(&k) {
 LL | |         None
@@ -100,7 +73,7 @@ LL | |     };
    | |_____^
    |
 help: consider using `m.entry(k)`
-  --> $DIR/entry.rs:48:5
+  --> $DIR/entry_unfixable.rs:35:5
    |
 LL | /     if m.contains_key(&k) {
 LL | |         None
@@ -111,7 +84,7 @@ LL | |     };
    | |_____^
 
 error: usage of `contains_key` followed by `insert` on a `BTreeMap`
-  --> $DIR/entry.rs:57:5
+  --> $DIR/entry_unfixable.rs:44:5
    |
 LL | /     if !m.contains_key(&k) {
 LL | |         foo();
@@ -122,7 +95,7 @@ LL | |     };
    | |_____^
    |
 help: consider using `m.entry(k)`
-  --> $DIR/entry.rs:57:5
+  --> $DIR/entry_unfixable.rs:44:5
    |
 LL | /     if !m.contains_key(&k) {
 LL | |         foo();
@@ -132,5 +105,5 @@ LL | |         None
 LL | |     };
    | |_____^
 
-error: aborting due to 7 previous errors
+error: aborting due to 5 previous errors
