+++ tests/ui/string_lit_as_bytes.fixed
@@ -14,7 +14,7 @@ fn str_lit_as_bytes() {
 
     let strify = stringify!(foobar).as_bytes();
 
-    let includestr = include_bytes!("entry.rs");
+    let includestr = include_bytes!("entry_unfixable.rs");
 }
 
 fn main() {}
